Install Foreman in an entrypoint
Since 3ce676a589c8449c33197508a5f8d510d5c09b50, we persist system gems
in `/bundle` which is supposed to be a Docker volume. It means that any
gems installed from the Dockerfile will be installed there and thus
overwritten when the volume is mount (and appear as not installed).

This commit stops installing `foreman` (the only system gem so far) from
the `Dockerfile`, but does it from the entrypoint (when the volume is
already mounted on `/bundle`).

I tried to configure `GEM_HOME` and `GEM_PATH` to include multiple paths
(_i.e._ `/usr/local/bundle/` for gems installed from the `Dockerfile`
and `/bundle/global` for gems installed globally from within the
container), but couldn't make it work in a way that would always work
(for all versions of Ruby, mostly).

Another option was to use UnionFS to merge `/usr/local/bundle` and
`/bundle/global` but it's more complex to setup (it's still an option
for the future, though).