Remove hardcoded Yarn and Node versions
They were used to be able to bump them and force a rebuild, but since
neither of those APT repositories maintain old versions of the package,
they are causing build failures as soon as a new version is available.

Thus considered a liability, they have been removed, and we will always
use the latest version or those packages.