Keep Bundler 1 for autoswitching
As part of a failed attempt of migrating people to Bundler 2 smoothly,
the Bundler team implemented a controversial (or, at least, "confusing")
feature where Bundler autoswitches to Bundler 1.x (if installed) when it
processes a `Gemfile.lock` with `BUNDLED WITH 1.x` (after
autoswitching to Bundler 1.x, it forcefully upgrades the
`Gemfile.lock` with `BUNDLED WITH 2.x`, which is another controversial
decision).

Anyway, this commit adds back Bundler 1.17 to avoid breaking this
autoswitching feature.

More information:
* https://eregon.me/blog/2020/01/13/a-migration-path-to-bundler2.html