Remove app binstubs from $PATH
We were including project binstubs (the project `bin/` directory,
actually) as highest priority entry in the `$PATH` env variable.

This may have been useful at some point, but I'm not sure why we did
that. Anyway, after many discussions with other people (including some
people who do that in their host environment), it appears that it should
not be done automatically. People who need/want this behaviour can
configure it, but it should definitely not be a default setting.

This commit removes `/app/bin` from the `$PATH` inside the Docker
containter.