Move foreman install to its own entrypoint
This will allow some images based on this one to disable automatic `bundle install` by overwriting `ENTRYPOINT` (to not include `bundler-wrapper`) without losing foreman too.

A typical use-case is an image for production, based on this one, and which would run `bundle install` and `yarn install` directly during the build (in `Dockerfile`).