Replace `bypass` entrypoint by an environment variable
The `bypass` entrypoint was a simple link to `gosu-entrypoint`. It was thus skipping all other entrypoints, even those that should be useful like `foreman-entrypoint`.

Since the `bypass` entrypoint was created to skip dependencies auto-installation, this commit replaces it with an environment variable (`DISABLE_AUTO_INSTALL_DEPS`) which can be set to `1` to skip both Bundler and Yarn entrypoints.