Fix a non-POSIX shell redirection in entrypoints
Since we switched entrypoint shebangs from `/bin/bash` to `/bin/sh` (under the hood: `dash` on Debian, `ash` on Alpine) in d70ebaa94fde0e608a489cd59abf4ba2cc9d4dc4, entrypoints must be POSIX compliant.

Unfortunately, the `&>` redirection is not POSIX-compliant and resulted in the left-hand command to be executed in background (without stdout/stdin redirection) because of the `&` character.

This commit uses a POSIX-compliant way to redirect both stdin and stdout to `/dev/null`.